+++ src/langcheck/metrics/_pairwise_text_quality_utils.py
@@ -67,8 +67,6 @@ def compute_pairwise_comparison_metric_values_with_consistency(
     metric_name: str,
     language: str,
     score_map: dict[str, float],
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Utility function to compute the pairwise metric values from the given
     Jinja template with the metric inputs. This function always enforces the
@@ -86,8 +84,6 @@ def compute_pairwise_comparison_metric_values_with_consistency(
         language: The language of the prompts. (e.g. "en")
         score_map: The mapping from the short assessment results
             (e.g. "Good") to the scores.
-        score_eval_client: The EvalClient instance used for the score evaluation.
-            If not provided, the scores will be computed using the `eval_client`.
 
     Returns:
         MetricValue: The metric values computed from the template.
@@ -98,24 +94,12 @@ def compute_pairwise_comparison_metric_values_with_consistency(
         for prompt_template_input in prompt_template_inputs
     ]
 
-    if score_eval_client:
-        explanations = eval_client.get_text_responses(
-            prompts=populated_prompts,
-        )
-        scores = score_eval_client.get_float_score(
-            metric_name=metric_name,
-            language=language,
-            unstructured_assessment_result=explanations,
-            score_map=score_map,
-        )
-
-    else:
-        scores, explanations = eval_client.get_score(
-            metric_name=metric_name,
-            language=language,
-            prompts=populated_prompts,
-            score_map=score_map,
-        )
+    scores, explanations = eval_client.get_score(
+        metric_name=metric_name,
+        language=language,
+        prompts=populated_prompts,
+        score_map=score_map,
+    )
 
     # Swap the generated outputs and sources to enforce consistency
     swapped_prompt_template_inputs = (
@@ -130,29 +114,14 @@ def compute_pairwise_comparison_metric_values_with_consistency(
         "[Swapped model outputs order] Intermediate assessments (1/2)"
     )
     score_tqdm = "[Swapped model outputs order] Calculating scores (2/2)"
-
-    if score_eval_client:
-        swapped_explanations = eval_client.get_text_responses(
-            prompts=swapped_prompts,
-            tqdm_description=intermediate_tqdm,
-        )
-        swapped_scores = score_eval_client.get_float_score(
-            metric_name=metric_name,
-            language=language,
-            unstructured_assessment_result=swapped_explanations,
-            score_map=score_map,
-            tqdm_description=score_tqdm,
-        )
-
-    else:
-        swapped_scores, swapped_explanations = eval_client.get_score(
-            metric_name=metric_name,
-            language=language,
-            prompts=swapped_prompts,
-            score_map=score_map,
-            intermediate_tqdm_description=intermediate_tqdm,
-            score_tqdm_description=score_tqdm,
-        )
+    swapped_scores, swapped_explanations = eval_client.get_score(
+        metric_name=metric_name,
+        language=language,
+        prompts=swapped_prompts,
+        score_map=score_map,
+        intermediate_tqdm_description=intermediate_tqdm,
+        score_tqdm_description=score_tqdm,
+    )
 
     # NOTE: The enforce_pairwise_comparison_consistency function assumes
     # that the score_map is symmetric, in the sense that swapping Model A
