+++ src/langcheck/metrics/en/query_based_text_quality.py
@@ -1,12 +1,8 @@
 from __future__ import annotations
 
-from langcheck.metrics.compute_metric_value import (
-    compute_metric_values_from_template,
-)
 from langcheck.metrics.eval_clients import EvalClient
 from langcheck.metrics.metric_inputs import get_metric_inputs
 from langcheck.metrics.metric_value import MetricValue
-from langcheck.metrics.prompts._utils import load_prompt_template
 
 LANG = "en"
 
@@ -15,8 +11,6 @@ def answer_relevance(
     generated_outputs: list[str] | str,
     prompts: list[str] | str,
     eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the relevance of generated outputs to the prompt. This metric
     takes on float values of either 0.0 (Not Relevant), 0.5 (Partially
@@ -33,11 +27,11 @@ def answer_relevance(
 
     metric_name = "answer_relevance"
 
-    answer_relevance_template = load_prompt_template(
+    answer_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=answer_relevance_template,
         metric_name=metric_name,
@@ -47,8 +41,6 @@ def answer_relevance(
             "Partially Relevant": 0.5,
             "Fully Relevant": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
 
 
@@ -56,8 +48,6 @@ def answer_safety(
     generated_outputs: list[str] | str,
     prompts: list[str] | str,
     eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the safety of generated outputs to the prompt. This metric
     takes on float values of either 0.0 (Unsafe), 0.5 (Partially Unsafe), or 1.0
@@ -73,11 +63,11 @@ def answer_safety(
 
     metric_name = "answer_safety"
 
-    answer_relevance_template = load_prompt_template(
+    answer_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=answer_relevance_template,
         metric_name=metric_name,
@@ -87,8 +77,6 @@ def answer_safety(
             "Partially Unsafe": 0.5,
             "Safe": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
 
 
@@ -96,8 +84,6 @@ def personal_data_leakage(
     generated_outputs: list[str] | str,
     prompts: list[str] | str,
     eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the personal data leakage of generated outputs to the prompt.
     This metric takes on float values of either 0.0 (Low Risk),
@@ -114,11 +100,11 @@ def personal_data_leakage(
 
     metric_name = "personal_data_leakage"
 
-    answer_relevance_template = load_prompt_template(
+    answer_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=answer_relevance_template,
         metric_name=metric_name,
@@ -128,8 +114,6 @@ def personal_data_leakage(
             "Medium Risk": 0.5,
             "High Risk": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
 
 
@@ -137,8 +121,6 @@ def hate_speech(
     generated_outputs: list[str] | str,
     prompts: list[str] | str,
     eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates whether hate speech is included in the generated outputs to
     the prompt.  This metric takes on float values of either 0.0 (Low Risk),
@@ -155,11 +137,11 @@ def hate_speech(
 
     metric_name = "hate_speech"
 
-    answer_relevance_template = load_prompt_template(
+    answer_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=answer_relevance_template,
         metric_name=metric_name,
@@ -169,8 +151,6 @@ def hate_speech(
             "Medium Risk": 0.5,
             "High Risk": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
 
 
@@ -178,8 +158,6 @@ def adult_content(
     generated_outputs: list[str] | str,
     prompts: list[str] | str,
     eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates whether adult content is included in the generated outputs to
     the prompt.  This metric takes on float values of either 0.0 (Low Risk),
@@ -196,11 +174,11 @@ def adult_content(
 
     metric_name = "adult_content"
 
-    answer_relevance_template = load_prompt_template(
+    answer_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=answer_relevance_template,
         metric_name=metric_name,
@@ -210,8 +188,6 @@ def adult_content(
             "Medium Risk": 0.5,
             "High Risk": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
 
 
@@ -219,8 +195,6 @@ def harmful_activity(
     generated_outputs: list[str] | str,
     prompts: list[str] | str,
     eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates whether the suggestion or instruction for harmful activity is
     included in the generated outputs to the prompt.  This metric takes on float
@@ -237,11 +211,11 @@ def harmful_activity(
 
     metric_name = "harmful_activity"
 
-    answer_relevance_template = load_prompt_template(
+    answer_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=answer_relevance_template,
         metric_name=metric_name,
@@ -251,6 +225,4 @@ def harmful_activity(
             "Medium Risk": 0.5,
             "High Risk": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
