+++ src/langcheck/metrics/en/source_based_text_quality.py
@@ -7,16 +7,12 @@ from transformers.models.auto.configuration_auto import AutoConfig
 from transformers.models.auto.modeling_auto import AutoModelForSeq2SeqLM
 from transformers.models.auto.tokenization_auto import AutoTokenizer
 
-from langcheck.metrics.compute_metric_value import (
-    compute_metric_values_from_template,
-)
 from langcheck.metrics.eval_clients import EvalClient
 from langcheck.metrics.metric_inputs import (
     get_metric_inputs,
     get_metric_inputs_with_required_lists,
 )
 from langcheck.metrics.metric_value import MetricValue
-from langcheck.metrics.prompts._utils import load_prompt_template
 from langcheck.utils.progress_bar import tqdm_wrapper
 
 _factual_consistency_model_path = "MingZhong/unieval-fact"
@@ -32,8 +28,6 @@ def factual_consistency(
     sources: list[str] | str,
     prompts: list[str] | str | None = None,
     eval_model: str | EvalClient = "local",
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the factual consistency between the generated outputs and
     the sources. This metric takes on float values between [0, 1], where 0
@@ -59,9 +53,6 @@ def factual_consistency(
             optional metadata and not used to calculate the metric.
         eval_model: The type of model to use ('local' or the EvalClient instance
             used for the evaluation). default 'local'
-        score_eval_client (Optional): The EvalClient instance used for the score
-            evaluation. If not provided, the scores will be computed using the
-            `eval_model`.
 
     Returns:
         An MetricValue object
@@ -92,7 +83,7 @@ def factual_consistency(
             "An EvalClient must be provided for non-local model types."
         )
 
-        factual_consistency_template = load_prompt_template(
+        factual_consistency_template = eval_model.load_prompt_template(
             language=LANG, metric_name=metric_name
         )
         factual_consistency_assessment_to_score = {
@@ -101,14 +92,12 @@ def factual_consistency(
             "Not Consistent": 0.0,
         }
 
-        return compute_metric_values_from_template(
+        return eval_model.compute_metric_values_from_template(
             metric_inputs=metric_inputs,
             template=factual_consistency_template,
             metric_name=metric_name,
             language=LANG,
             score_map=factual_consistency_assessment_to_score,
-            eval_client=eval_model,
-            score_eval_client=score_eval_client,
         )
 
 
@@ -172,8 +161,8 @@ def _factual_consistency_local(
         )
         _factual_consistency_model.eval()
 
-    pos_id = _factual_consistency_tokenizer("Yes")["input_ids"][0]  # type: ignore
-    neg_id = _factual_consistency_tokenizer("No")["input_ids"][0]  # type: ignore
+    pos_id = _factual_consistency_tokenizer("Yes")["input_ids"][0]
+    neg_id = _factual_consistency_tokenizer("No")["input_ids"][0]
     softmax = nn.Softmax(dim=1)
 
     model_input_list = []
@@ -208,7 +197,7 @@ def _factual_consistency_local(
             )
             inputs_tokens = encoded_inputs["input_ids"]
             inputs_mask = encoded_inputs["attention_mask"]
-            targets_tokens = encoded_targets["input_ids"][:, 0].unsqueeze(-1)  # type: ignore
+            targets_tokens = encoded_targets["input_ids"][:, 0].unsqueeze(-1)
 
             outputs = _factual_consistency_model(
                 input_ids=inputs_tokens,
@@ -235,11 +224,7 @@ def _factual_consistency_local(
 
 
 def context_relevance(
-    sources: list[str] | str,
-    prompts: list[str] | str,
-    eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
+    sources: list[str] | str, prompts: list[str] | str, eval_model: EvalClient
 ) -> MetricValue[float | None]:
     """Calculates the relevance of the sources to the prompts. This metric takes
     on float values between [0, 1], where 0 means that the source text is not at
@@ -252,9 +237,6 @@ def context_relevance(
         sources: The source text(s), one string per prompt
         prompts: The prompt(s)
         eval_model: The EvalClient instance used for the evaluation
-        score_eval_client (Optional): The EvalClient instance used for the score
-            evaluation. If not provided, the scores will be computed using the
-            `eval_model`.
     """
     metric_inputs = get_metric_inputs(
         prompts=prompts,
@@ -263,7 +245,7 @@ def context_relevance(
     )
 
     metric_name = "context_relevance"
-    context_relevance_template = load_prompt_template(
+    context_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
@@ -273,12 +255,10 @@ def context_relevance(
         "Not Relevant": 0.0,
     }
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=context_relevance_template,
         metric_name=metric_name,
         language=LANG,
         score_map=context_relevance_assessment_to_score,
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
