+++ src/langcheck/metrics/eval_clients/_base.py
@@ -4,6 +4,10 @@ from typing import Union
 
 from jinja2 import Template
 
+from langcheck.metrics.metric_inputs import MetricInputs
+from langcheck.metrics.metric_value import MetricValue
+
+from ..prompts._utils import get_template
 from ..scorer._base import BaseSimilarityScorer
 
 TokenLogProb = dict[str, Union[str, float]]
@@ -16,6 +20,31 @@ class EvalClient:
     Most metrics that use external APIs such as OpenAI API call the methods
     defined in this class to compute the metric values.
     """
+
+    def load_prompt_template(
+        self,
+        language: str,
+        metric_name: str,
+        eval_prompt_version: str | None = None,
+    ) -> Template:
+        """
+        Gets a Jinja template from the specified language, eval client, metric
+        name, and (optionally) eval prompt version.
+
+        Args:
+            language (str): The language of the template.
+            metric_name (str): The name of the metric.
+            eval_prompt_version (str | None): The version of the eval prompt.
+                If None, the default version is used.
+
+        Returns:
+            Template: The Jinja template.
+        """
+        if eval_prompt_version is None:
+            return get_template(f"{language}/metrics/{metric_name}.j2")
+        return get_template(
+            f"{language}/metrics/{metric_name}_{eval_prompt_version}.j2"
+        )
 
     def get_text_responses(
         self,
@@ -149,6 +178,53 @@ class EvalClient:
         TODO: Intergrate scorer/ with eval_clients/
         """
         raise NotImplementedError
+
+    def compute_metric_values_from_template(
+        self,
+        metric_inputs: MetricInputs,
+        template: Template,
+        metric_name: str,
+        language: str,
+        score_map: dict[str, float],
+    ) -> MetricValue[float | None]:
+        """Compute the metric values from the given Jinja template with the
+        metric inputs. This function assumes that the template parameters are
+        already validated and the template is ready to be rendered.
+
+        Args:
+            metric_inputs: The metric inputs that contain the prompts,
+                generated outputs, reference outputs... etc.
+            template: The Jinja template that is ready to be rendered.
+            enforce_pairwise_consistency: Whether to enforce pairwise
+                consistency when computing the metric values.
+            metric_name: The name of the metric to be used. (e.g. "toxicity")
+            language: The language of the prompts. (e.g. "en")
+            score_map: The mapping from the short assessment results
+                (e.g. "Good") to the scores.
+
+        Returns:
+            MetricValue: The metric values computed from the template.
+        """
+        prompt_template_inputs = metric_inputs.get_inputs_for_prompt_template()
+        populated_prompts = [
+            template.render(prompt_template_input)
+            for prompt_template_input in prompt_template_inputs
+        ]
+
+        scores, explanations = self.get_score(
+            metric_name=metric_name,
+            language=language,
+            prompts=populated_prompts,
+            score_map=score_map,
+        )
+
+        return MetricValue(
+            metric_name=metric_name,
+            metric_inputs=metric_inputs,
+            explanations=explanations,
+            metric_values=scores,
+            language=language,
+        )
 
     def repeat_requests_from_template(
         self,
