+++ src/langcheck/metrics/ja/pairwise_text_quality.py
@@ -3,13 +3,9 @@ from __future__ import annotations
 from langcheck.metrics._pairwise_text_quality_utils import (
     compute_pairwise_comparison_metric_values_with_consistency,
 )
-from langcheck.metrics.compute_metric_value import (
-    compute_metric_values_from_template,
-)
 from langcheck.metrics.eval_clients import EvalClient
 from langcheck.metrics.metric_inputs import get_metric_inputs
 from langcheck.metrics.metric_value import MetricValue
-from langcheck.metrics.prompts._utils import load_prompt_template
 
 
 def pairwise_comparison(
@@ -21,8 +17,6 @@ def pairwise_comparison(
     reference_outputs: list[str] | str | None = None,
     enforce_consistency: bool = True,
     eval_model: EvalClient | None = None,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the pairwise comparison metric. This metric takes on float
     values of either 0.0 (Response A is better), 0.5 (Tie), or 1.0 (Response B
@@ -47,9 +41,6 @@ def pairwise_comparison(
             marked as Optional so that it can follow the above arguments that
             have default values (for consistency with the other metrics), but
             this is in fact a required argument.
-        score_eval_client (Optional): The EvalClient instance used for the score
-            evaluation. If not provided, the scores will be computed using the
-            `eval_model`.
 
     Returns:
         An MetricValue object
@@ -73,7 +64,7 @@ def pairwise_comparison(
     }
     metric_name = "pairwise_comparison"
     language = "ja"
-    pairwise_comparison_template = load_prompt_template(
+    pairwise_comparison_template = eval_model.load_prompt_template(
         language=language, metric_name=metric_name
     )
 
@@ -87,12 +78,10 @@ def pairwise_comparison(
             score_map=pairwise_comparison_assessment_to_score,
         )
     else:
-        return compute_metric_values_from_template(
+        return eval_model.compute_metric_values_from_template(
             metric_inputs=metric_inputs,
             template=pairwise_comparison_template,
             metric_name=metric_name,
             language=language,
             score_map=pairwise_comparison_assessment_to_score,
-            eval_client=eval_model,
-            score_eval_client=score_eval_client,
         )
