+++ src/langcheck/metrics/ja/reference_free_text_quality.py
@@ -2,16 +2,12 @@ from __future__ import annotations
 
 import regex as re
 
-from langcheck.metrics.compute_metric_value import (
-    compute_metric_values_from_template,
-)
 from langcheck.metrics.eval_clients import EvalClient
 from langcheck.metrics.metric_inputs import (
     get_metric_inputs,
     get_metric_inputs_with_required_lists,
 )
 from langcheck.metrics.metric_value import MetricValue
-from langcheck.metrics.prompts._utils import load_prompt_template
 from langcheck.metrics.scorer.hf_models import (
     AutoModelForSequenceClassificationScorer,
 )
@@ -25,8 +21,6 @@ def sentiment(
     prompts: list[str] | str | None = None,
     eval_model: str | EvalClient = "local",
     local_overflow_strategy: str = "truncate",
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the sentiment scores of generated outputs. This metric takes
     on float values between [0, 1], where 0 is negative sentiment and 1 is
@@ -59,8 +53,6 @@ def sentiment(
             will be assigned a score of None. If 'truncate', the outputs that
             are too long will be truncated. If 'raise', an error will be raised
             when the outputs are too long. The default value is 'nullify'.
-        score_eval_client (Optional): The EvalClient instance used for the score evaluation.
-            If not provided, the scores will be computed using the `eval_model`.
 
     Returns:
         An :class:`~langcheck.metrics.metric_value.MetricValue` object
@@ -87,7 +79,7 @@ def sentiment(
             "An EvalClient must be provided for non-local model types."
         )
 
-        sentiment_template = load_prompt_template(
+        sentiment_template = eval_model.load_prompt_template(
             language=LANG, metric_name=metric_name
         )
 
@@ -97,14 +89,12 @@ def sentiment(
             "Negative": 0.0,
         }
 
-        return compute_metric_values_from_template(
+        return eval_model.compute_metric_values_from_template(
             metric_inputs=metric_inputs,
             template=sentiment_template,
             metric_name=metric_name,
             language=LANG,
             score_map=sentiment_assessment_to_score,
-            eval_client=eval_model,
-            score_eval_client=score_eval_client,
         )
 
 
@@ -145,8 +135,6 @@ def toxicity(
     eval_model: str | EvalClient = "local",
     local_overflow_strategy: str = "truncate",
     eval_prompt_version: str = "v2",
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the toxicity scores of generated outputs. This metric takes on
     float values between [0, 1], where 0 is low toxicity and 1 is high toxicity.
@@ -185,8 +173,6 @@ def toxicity(
             when the outputs are too long. The default value is 'nullify'.
         eval_prompt_version: The version of the eval prompt to use when the
             EvalClient is used. The default version is 'v2' (latest).
-        score_eval_client (Optional): The EvalClient instance used for the score evaluation.
-            If not provided, the scores will be computed using the `eval_model`.
 
     Returns:
         An :class:`~langcheck.metrics.metric_value.MetricValue` object
@@ -231,20 +217,18 @@ def toxicity(
             f"Invalid eval_prompt_version: {eval_prompt_version}. The valid versions are {list(toxicity_assessment_to_score.keys())}."
         )
 
-        toxicity_template = load_prompt_template(
+        toxicity_template = eval_model.load_prompt_template(
             language=LANG,
             metric_name=metric_name,
             eval_prompt_version=eval_prompt_version,
         )
 
-        return compute_metric_values_from_template(
+        return eval_model.compute_metric_values_from_template(
             metric_inputs=metric_inputs,
             template=toxicity_template,
             metric_name=metric_name,
             language=LANG,
             score_map=toxicity_assessment_to_score[eval_prompt_version],
-            eval_client=eval_model,
-            score_eval_client=score_eval_client,
         )
 
 
@@ -283,8 +267,6 @@ def fluency(
     prompts: list[str] | str | None = None,
     eval_model: str | EvalClient = "local",
     local_overflow_strategy: str = "truncate",
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the fluency scores of generated outputs. This metric takes on
     float values between [0, 1], where 0 is low fluency and 1 is high fluency.
@@ -320,8 +302,6 @@ def fluency(
             will be assigned a score of None. If 'truncate', the outputs that
             are too long will be truncated. If 'raise', an error will be raised
             when the outputs are too long. The default value is 'nullify'.
-        score_eval_client (Optional): The EvalClient instance used for the score evaluation.
-            If not provided, the scores will be computed using the `eval_model`.
 
     Returns:
         An :class:`~langcheck.metrics.metric_value.MetricValue` object
@@ -348,7 +328,7 @@ def fluency(
             "An EvalClient must be provided for non-local model types."
         )
 
-        fluency_template = load_prompt_template(
+        fluency_template = eval_model.load_prompt_template(
             language=LANG, metric_name=metric_name
         )
 
@@ -358,14 +338,12 @@ def fluency(
             "Good": 1.0,
         }
 
-        return compute_metric_values_from_template(
+        return eval_model.compute_metric_values_from_template(
             metric_inputs=metric_inputs,
             template=fluency_template,
             metric_name=metric_name,
             language=LANG,
             score_map=fluency_assessment_to_score,
-            eval_client=eval_model,
-            score_eval_client=score_eval_client,
         )
 
 
@@ -492,8 +470,6 @@ def tateishi_ono_yamada_reading_ease(
 def jailbreak_prompt(
     prompts: list[str] | str,
     eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates whether jailbreak techniques are included in the prompts.
     This metric takes on float values of either 0.0 (Low Risk),
@@ -508,11 +484,11 @@ def jailbreak_prompt(
     )
 
     metric_name = "jailbreak_prompt"
-    jailbreak_prompt_template = load_prompt_template(
+    jailbreak_prompt_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=jailbreak_prompt_template,
         metric_name=metric_name,
@@ -522,8 +498,6 @@ def jailbreak_prompt(
             "Medium Risk": 0.5,
             "High Risk": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
 
 
@@ -532,8 +506,6 @@ def prompt_leakage(
     system_prompts: list[str] | str,
     eval_model: EvalClient,
     eval_prompt_version: str = "v2",
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the severity of prompt leakage in the generated outputs.
     This metric takes on float values of either 0.0 (Low Risk),
@@ -559,13 +531,13 @@ def prompt_leakage(
 
     metric_name = "prompt_leakage"
 
-    prompt_leakage_template = load_prompt_template(
+    prompt_leakage_template = eval_model.load_prompt_template(
         language=LANG,
         metric_name=metric_name,
         eval_prompt_version=eval_prompt_version,
     )
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=prompt_leakage_template,
         metric_name=metric_name,
@@ -575,6 +547,4 @@ def prompt_leakage(
             "Medium Risk": 0.5,
             "High Risk": 1.0,
         },
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
