+++ src/langcheck/metrics/ja/source_based_text_quality.py
@@ -5,9 +5,6 @@ from typing import cast
 from transformers.pipelines import pipeline
 from transformers.pipelines.base import Pipeline
 
-from langcheck.metrics.compute_metric_value import (
-    compute_metric_values_from_template,
-)
 from langcheck.metrics.en.source_based_text_quality import (
     factual_consistency as en_factual_consistency,
 )
@@ -17,7 +14,6 @@ from langcheck.metrics.metric_inputs import (
     get_metric_inputs_with_required_lists,
 )
 from langcheck.metrics.metric_value import MetricValue
-from langcheck.metrics.prompts._utils import load_prompt_template
 from langcheck.utils.progress_bar import tqdm_wrapper
 
 _factual_consistency_translation_model_path = "Helsinki-NLP/opus-mt-ja-en"
@@ -31,8 +27,6 @@ def factual_consistency(
     sources: list[str] | str,
     prompts: list[str] | str | None = None,
     eval_model: str | EvalClient = "local",
-    *,
-    score_eval_client: EvalClient | None = None,
 ) -> MetricValue[float | None]:
     """Calculates the factual consistency between the generated outputs and
     the sources. This metric takes on float values between [0, 1], where 0
@@ -61,10 +55,8 @@ def factual_consistency(
         sources: The source text(s), one string per generated output
         prompts: The prompts used to generate the output(s). Prompts are
             optional metadata and not used to calculate the metric.
-        eval_model: The type of mode to use ('local' or the EvalClient instance
+        eval_model: The type of model to use ('local' or the EvalClient instance
             used for the evaluation). default 'local'
-        score_eval_client (Optional): The EvalClient instance used for the score evaluation.
-            If not provided, the scores will be computed using the `eval_model`.
 
     Returns:
         An MetricValue object
@@ -95,7 +87,7 @@ def factual_consistency(
             "An EvalClient must be provided for non-local model types."
         )
 
-        factual_consistency_template = load_prompt_template(
+        factual_consistency_template = eval_model.load_prompt_template(
             language=LANG, metric_name=metric_name
         )
         factual_consistency_assessment_to_score = {
@@ -104,14 +96,12 @@ def factual_consistency(
             "Not Consistent": 0.0,
         }
 
-        return compute_metric_values_from_template(
+        return eval_model.compute_metric_values_from_template(
             metric_inputs=metric_inputs,
             template=factual_consistency_template,
             metric_name=metric_name,
             language=LANG,
             score_map=factual_consistency_assessment_to_score,
-            eval_client=eval_model,
-            score_eval_client=score_eval_client,
         )
 
 
@@ -192,11 +182,7 @@ def _factual_consistency_local(
 
 
 def context_relevance(
-    sources: list[str] | str,
-    prompts: list[str] | str,
-    eval_model: EvalClient,
-    *,
-    score_eval_client: EvalClient | None = None,
+    sources: list[str] | str, prompts: list[str] | str, eval_model: EvalClient
 ) -> MetricValue[float | None]:
     """Calculates the relevance of the sources to the prompts. This metric takes
     on float values between [0, 1], where 0 means that the source text is not at
@@ -209,8 +195,6 @@ def context_relevance(
         sources: The source text(s), one string per prompt
         prompts: The prompt(s)
         eval_model: The EvalClient instance used for the evaluation
-        score_eval_client (Optional): The EvalClient instance used for the score evaluation.
-            If not provided, the scores will be computed using the `eval_model`.
     """
     metric_inputs = get_metric_inputs(
         prompts=prompts,
@@ -219,7 +203,7 @@ def context_relevance(
     )
 
     metric_name = "context_relevance"
-    context_relevance_template = load_prompt_template(
+    context_relevance_template = eval_model.load_prompt_template(
         language=LANG, metric_name=metric_name
     )
 
@@ -229,12 +213,10 @@ def context_relevance(
         "Not Relevant": 0.0,
     }
 
-    return compute_metric_values_from_template(
+    return eval_model.compute_metric_values_from_template(
         metric_inputs=metric_inputs,
         template=context_relevance_template,
         metric_name=metric_name,
         language=LANG,
         score_map=context_relevance_assessment_to_score,
-        eval_client=eval_model,
-        score_eval_client=score_eval_client,
     )
